chore: move http.Handler to http.go
