+++ pkg/limits/frontend/frontend.go
@@ -7,10 +7,8 @@ package frontend
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"net/http"
 	"time"
 
 	"github.com/go-kit/log"
@@ -18,12 +16,10 @@ import (
 	"github.com/grafana/dskit/limiter"
 	"github.com/grafana/dskit/ring"
 	"github.com/grafana/dskit/services"
-	"github.com/grafana/dskit/user"
 	"github.com/prometheus/client_golang/prometheus"
 
 	limits_client "github.com/grafana/loki/v3/pkg/limits/client"
 	"github.com/grafana/loki/v3/pkg/logproto"
-	"github.com/grafana/loki/v3/pkg/util"
 	util_log "github.com/grafana/loki/v3/pkg/util/log"
 )
 
@@ -108,14 +104,6 @@ func New(cfg Config, ringName string, limitsRing ring.ReadRing, limits Limits, l
 	return f, nil
 }
 
-// Flush implements ring.FlushTransferer. It transfers state to another ingest limits frontend instance.
-func (f *Frontend) Flush() {}
-
-// TransferOut implements ring.FlushTransferer. It transfers state to another ingest limits frontend instance.
-func (f *Frontend) TransferOut(_ context.Context) error {
-	return nil
-}
-
 // starting implements services.Service.
 func (f *Frontend) starting(ctx context.Context) (err error) {
 	defer func() {
@@ -170,58 +158,12 @@ func (f *Frontend) CheckReady(ctx context.Context) error {
 	return nil
 }
 
-type exceedsLimitsRequest struct {
-	TenantID     string   `json:"tenantID"`
-	StreamHashes []uint64 `json:"streamHashes"`
-}
+// Flush implements ring.FlushTransferer. It transfers state to another ingest
+// limits frontend instance.
+func (f *Frontend) Flush() {}
 
-type exceedsLimitsResponse struct {
-	RejectedStreams []*logproto.RejectedStream `json:"rejectedStreams,omitempty"`
-}
-
-// ServeHTTP implements http.Handler.
-func (f *Frontend) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	var req exceedsLimitsRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		level.Error(f.logger).Log("msg", "error unmarshalling request body", "err", err)
-		http.Error(w, "error unmarshalling request body", http.StatusBadRequest)
-		return
-	}
-
-	if req.TenantID == "" {
-		http.Error(w, "tenantID is required", http.StatusBadRequest)
-		return
-	}
-
-	// Convert request to protobuf format
-	protoReq := &logproto.ExceedsLimitsRequest{
-		Tenant:  req.TenantID,
-		Streams: make([]*logproto.StreamMetadata, len(req.StreamHashes)),
-	}
-	for i, hash := range req.StreamHashes {
-		protoReq.Streams[i] = &logproto.StreamMetadata{
-			StreamHash: hash,
-		}
-	}
-
-	ctx, err := user.InjectIntoGRPCRequest(user.InjectOrgID(r.Context(), req.TenantID))
-	if err != nil {
-		http.Error(w, "failed to inject org ID", http.StatusInternalServerError)
-		return
-	}
-
-	// Call the service
-	resp, err := f.limits.ExceedsLimits(ctx, protoReq)
-	if err != nil {
-		level.Error(f.logger).Log("msg", "error checking limits", "err", err)
-		http.Error(w, "error checking limits", http.StatusInternalServerError)
-		return
-	}
-
-	// Convert response to JSON format
-	jsonResp := exceedsLimitsResponse{
-		RejectedStreams: resp.RejectedStreams,
-	}
-
-	util.WriteJSONResponse(w, jsonResp)
+// TransferOut implements ring.FlushTransferer. It transfers state to another
+// ingest limits frontend instance.
+func (f *Frontend) TransferOut(_ context.Context) error {
+	return nil
 }
